refresh

+++ decompiled/sub_439420.c
@@ -3,7 +3,7 @@ int __cdecl sub_439420(LPVOID lpMem)
 {
   if ( !lpMem )
     return -2;
-  if ( *((_BYTE *)lpMem + 92) == 119 && !sub_439110(lpMem, 4) )
+  if ( *((_BYTE *)lpMem + 92) == 119 && !sub_439110((int)lpMem, 4) )
   {
     sub_439470(*((FILE **)lpMem + 16), *((_DWORD *)lpMem + 19));
     sub_439470(*((FILE **)lpMem + 16), *((_DWORD *)lpMem + 2));
